requirements.txt
# Core dependencies
streamlit>=1.31.0,<2.0
beautifulsoup4>=4.12.2
requests>=2.31.0
altair>=5.0.0,<6.0
anyio>=4.0.0,<5.0
appdirs
attrs>=25.0.0,<26.0
backoff>=2.2.1,<3.0
blinker>=1.9.0,<2.0
boilerpy3>=1.0.7,<2.0
cachetools>=5.5.0,<6.0
cattrs>=24.0.0,<25.0
certifi>=2025.0.0,<2026.0
charset-normalizer>=3.4.0,<4.0
click>=8.1.0,<9.0
colorama>=0.4.6,<1.0
distro>=1.9.0,<2.0
docopt>=0.6.2,<1.0
Events>=0.5,<1.0
farm-haystack>=1.24.0,<2.0
filelock>=3.18.0,<4.0
fsspec>=2025.0.0,<2026.0
gitdb>=4.0.0,<5.0
GitPython>=3.1.0,<4.0
h11>=0.14.0,<1.0
httpcore>=1.0.0,<2.0
httpx>=0.28.0,<1.0
huggingface-hub>=0.29.0,<1.0
idna>=3.10,<4.0
inflect>=7.5.0,<8.0
Jinja2>=3.1.0,<4.0
joblib>=1.4.0,<2.0
jsonschema>=4.23.0,<5.0
jsonschema-specifications>=2024.0.0,<2025.0
lazy-imports>=0.3.0,<1.0
MarkupSafe>=3.0.0,<4.0
monotonic>=1.6,<2.0
more-itertools>=10.0.0,<11.0
narwhals>=1.32.0,<2.0
networkx>=3.4.0,<4.0
nltk>=3.8.1,<4.0
num2words>=0.5.0,<1.0
numpy>=2.0.0,<3.0
packaging>=24.0,<25.0
pandas>=2.2.0,<3.0
pillow>=11.0.0,<12.0
platformdirs>=4.3.0,<5.0
posthog>=3.21.0,<4.0
prompthub-py>=4.0.0,<5.0
protobuf>=5.29.0,<6.0
pyarrow>=19.0.0,<20.0
pydantic>=1.10.0,<2.0  # Required for farm-haystack
pydeck>=0.9.0,<1.0
python-dateutil>=2.9.0,<3.0
pytz>=2025.0.0,<2026.0
PyYAML>=6.0.0,<7.0
quantulum3>=0.9.0,<1.0
rank-bm25>=0.2.0,<1.0
referencing>=0.36.0,<1.0
regex>=2024.0.0,<2025.0
requests-cache>=0.9.0,<1.0
rpds-py>=0.23.0,<1.0
safetensors>=0.5.0,<1.0
scikit-learn>=1.6.0,<2.0
scipy>=1.15.0,<2.0
six>=1.17.0,<2.0
smmap>=5.0.0,<6.0
sniffio>=1.3.0,<2.0
soupsieve>=2.6,<3.0
sseclient-py>=1.8.0,<2.0
tenacity>=9.0.0,<10.0
threadpoolctl>=3.6.0,<4.0
tiktoken>=0.9.0,<1.0
tokenizers>=0.21.0,<1.0
toml>=0.10.0,<1.0
tornado>=6.4.0,<7.0
tqdm>=4.67.0,<5.0
transformers>=4.50.0,<5.0
typeguard>=4.4.0,<5.0
typing_extensions>=4.12.0,<5.0
tzdata>=2025.0.0,<2026.0
url-normalize>=1.4.0,<2.0
urllib3>=2.3.0,<3.0
watchdog>=6.0.0,<7.0

# Optional dependencies
poppler-utils  # For PDF processing
python-magic  # For file processing
youtube-transcript-api>=0.6.0,<1.0  # For YouTube transcript extraction